EnvironmentConfig: Substitute Vraibles functionality moved to IShell scope
