+++ Source/WPEFramework/PluginServer.cpp
@@ -625,6 +625,11 @@ namespace PluginHost
             Core::Directory(persistentPath.Name().c_str()).Create();
         }
 
+        if (configuration.Environments.IsSet() == true) {
+            PluginHost::Environment environment;
+            environment.Set(_config, configuration.Environments);
+        }
+
         Core::JSON::ArrayType<Plugin::Config>::Iterator index = configuration.Plugins.Elements();
 
         // First register all services, than if we got them, start "activating what is required.
